requirements.txt
Flask==2.3.2
Flask-Mail==0.9.1
mysql-connector-python==8.0.33
pytest==7.4.0
flake8==6.1.0
python-dotenv==1.0.0
gunicorn==20.1.0
